% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignSeries.R
\name{alignSeries}
\alias{alignSeries}
\title{Align series to a reference}
\usage{
alignSeries(
  x,
  ref = c("median", "mean", "more options in documentation")[1],
  threshold = 0.6,
  shift = TRUE,
  padding = c("zeroes", "circular", "sampling")[1],
  from = as.integer(length(x) * 14/15):length(x),
  ...
)
}
\arguments{
\item{x}{numeric vector or matrix, series in rows.}

\item{ref}{logical, character, numeric or function. Specifies the reference
series for alignment. It is interpreted according to type:\itemize{
\item Numeric vector of the same length as the input series: literal reference
\item A number \code{<= dim(x)[1]}: row index of the reference series
\item logical of length \code{== dim(x)[2])}: logical filter that selects a
single row of \code{x} to be used as reference
\item function: the function is called with \code{x} as argument, the
returned value is used as reference
\item character: "median" (default) or "mean", use the median ir mean series
as reference.}}

\item{threshold}{numeric, the cross-correlation must be above this
threshold for the series to be aligned, otherwise it is left unshifted.
Increase this value if you know there are very big shifts. Decrease it if
you observe over-shifting.}

\item{shift}{logical. If TRUE (default), returns the aligned series. If FALSE,
returns the shifts that align the series.}

\item{padding}{character, the method to be used to fill the empty extremes of
the shifted series. See \code{\link{pad}} for details. Default: "zeroes".}

\item{from}{numeric, optional. Filter for the region to be used for padding
after the series is shifted in the "sampling" method. See \code{\link{pad}}
for details. Default: last 1/15th points.}

\item{...}{additional arguments to \code{\link[stats]{ccf}}}
}
\value{
either a numeric vector of shifts or a matrix with the shifted series
in its rows
@import methods
}
\description{
Invokes \code{\link[stats]{ccf}} to compute the cross-correlations
between the series to be aligned and the reference series and extracts the
shifts that maximize the cross-correlations
}
