% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineCorrection.R
\name{baselineCorrection}
\alias{baselineCorrection}
\title{Baseline Correction}
\usage{
baselineCorrection(Y)
}
\arguments{
\item{Y}{matrix, series in rows}
}
\value{
a matrix of same size with each row being corrected
}
\description{
Baseline Correction
}
