% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSpectra.R
\name{calibrateSignal}
\alias{calibrateSignal}
\title{Signal calibration}
\usage{
calibrateSignal(
  ppm,
  y,
  signal,
  maxShift = 1/3,
  threshold = 0.2,
  rOref = signalDomain(signal, 30),
  ...
)
}
\arguments{
\item{ppm}{numeric, chemical shift scale of the spectrum}

\item{y}{numeric, spectrum intensities.}

\item{signal}{\code{\linkS4class{NMRSignal1D}} to be calibrated}

\item{maxShift}{numeric, the maximum allowed shift as a fraction of the rOref,
see \code{\link[stats]{ccf}}.}

\item{threshold}{numeric, minimum cross-correlation required for alignment.
See \code{\link{alignSeries}}.}

\item{rOref}{numeric, optional, limits of the Region of Reference within
which the reference signal will be calibrated. The default encompasses a
region 30 linewidths around the external peaks.}

\item{...}{additional arguments for \code{\link{alignSeries}}.}
}
\value{
a \code{\linkS4class{NMRSignal1D}} with peak maxima shifted to
align with the reference spectrum
}
\description{
Calibrates a \code{\linkS4class{NMRSignal1D}} to the experimental spectrum
}
\details{
Interpolates the signal with \code{\link{signalToY}} and aligns the
resulting trace to the spectrum in the rOref using \code{\link{alignSeries}}.
This is useful e.g. for fine-tuning a quantification model.
}
