% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSpectra.R
\name{calibrateSpectra}
\alias{calibrateSpectra}
\title{Spectra calibration
Calibrates spectra by aligning a reference signal}
\usage{
calibrateSpectra(x, Y, RoRef, matrix, ...)
}
\arguments{
\item{x}{numeric, spectra scale e.g. ppm}

\item{Y}{matrix, intensities, spectra in rows}

\item{RoRef}{numeric or logical, optional. Specification of the Region of
Reference corresponding to the reference signal for calibration. It may be
specified as a vector of x coordinates, as a length 2 vector declaring the
limits of the region, or as logical filter for x. Either \code{RoRef} or \code{matrix}
must be provided. If \code{RoRef} is not provided, the standard reference peak
for the given \code{matrix} will be used as Region of Reference.}

\item{matrix}{character, optional. The sample matrix; either 'urine', 'plasma' or
'other'. Either \code{matrix} or \code{RoRef} must be specified.}

\item{...}{additional arguments for \code{alignSeries} (see details)}
}
\value{
calibrated spectra matrix
}
\description{
Spectra calibration
Calibrates spectra by aligning a reference signal
}
\details{
crops the Region of Reference of the spectra matrix and passes it
to \code{alignSeries}, which calculates the shifts that align the spectra on the
Region of Reference. Then, it shifts the full spectra by the
corresponding amounts.
}
