% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{crop}
\alias{crop}
\title{Create a filter to crop a spectral region}
\usage{
crop(ppm, start = -Inf, end = Inf, roi)
}
\arguments{
\item{ppm}{numeric, chemical shift scale e.g. ppm}

\item{start, }{numeric, optional. Lower limit of the \code{roi}.}

\item{end, }{numeric, optional. Upper limit of the \code{roi}.}

\item{roi}{numeric, optional. Upper and lower limit of the Region of Interest
to be cropped.}
}
\value{
logic, a filter for the elements of \code{ppm} within the \code{roi}.
}
\description{
Quality of life function to spare you a few uncomfortable key strokes and
some neural pulses. Questionable value.
}
\details{
Either \code{start}, \code{end} or \code{roi} is required. Argument
\code{roi} has priority. If \code{start} is given but not \code{end}, the
lower limit is effectively set to \code{min(ppm)}. If \code{end} is given but
not \code{start}, the upper limit is effectively set to \code{max(ppm)}.
}
