% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gaussian}
\alias{gaussian}
\title{Interpolates a gaussian with the given parameters on the given points.}
\usage{
gaussian(x, mean = 0, max = 1, fwhm = 1)
}
\arguments{
\item{x}{numeric, points to be interpolated}

\item{mean}{numeric, mean(s) of the gaussian(s)}

\item{max}{numeric, maxima of the gaussian(s)}

\item{fwhm}{numeric, full-width-at-half-max(s) of the gaussians(s)}
}
\value{
numeric, vector of interpolated values
}
\description{
Interpolates a gaussian with the given parameters on the given points.
}
\examples{
gaussian(1:100,max=1,mean=0)
}
