% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getI}
\alias{getI}
\title{Get the index of a chemical shift}
\usage{
getI(ppm, ...)
}
\arguments{
\item{ppm, }{numeric, chemical shift scale}

\item{..., }{numeric, query chemical shift values}
}
\value{
integer, the indices of the elements of ppm that are closest to the queried values
}
\description{
Quality of life function to spare you a few key strokes and some neural pulses.
}
\details{
Seeks the element(s) of the chemical shift scale closest to the given value(s).
Useful e.g. to get the approximate intensity of spectra at a given c. shift.
}
