% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integral.R
\name{integral}
\alias{integral}
\alias{integral,NMRSignal1D,character-method}
\alias{integral,NMRSignalModel,character-method}
\alias{integral,NMRSignal1D,function-method}
\alias{integral,NMRSignalModel,function-method}
\alias{integral,list,character-method}
\alias{integral,list,function-method}
\title{Integrate NMR signals}
\usage{
integral(nmr, method, ...)

\S4method{integral}{NMRSignal1D,character}(nmr, method, delta = 0.00023, offset = 3)

\S4method{integral}{NMRSignalModel,character}(nmr, method)

\S4method{integral}{NMRSignal1D,`function`}(nmr, method)

\S4method{integral}{NMRSignalModel,`function`}(nmr, method)

\S4method{integral}{list,character}(nmr, method)

\S4method{integral}{list,`function`}(nmr, method)
}
\arguments{
\item{nmr}{a \code{\linkS4class{NMRSignalModel}}, \code{\linkS4class{NMRSignal1D}}
or a list with named elements \emph{x} and \emph{y}}

\item{method}{character or function, integration method to be used.
If a string, it is interpreted as a specifier for an implemented
method:
\itemize{
\item "rect", Riemann sum approximation;
\item "fwhm", full-width at half-maximum approximation (not valid for list)
}
If a function, this function is used to compute the integrals with \code{nmr}
as argument (in the case of a list the elements, x and y are passed as separate
arguments.}

\item{...}{additional arguments}

\item{delta}{numeric, the chemical shift gap between interpolated points.
Though the default should work in most scenarios, you may want to adjust it to
the resolution of your experimental spectra.}

\item{offset}{numeric, a factor of the signal's full-width at half-maximum
(fwhm) used to determine the domain of integration: the signal is integrated
in a domain that offsets its outer peaks by \code{offset * fwhm}.}
}
\value{
numeric, the signal's integral

numeric, a vector of integrals for each signal in the NMRSignalModel

numeric, the integral of the data series
}
\description{
Integrate NMR signals
}
\details{
In the "rect" method, signals are constrained to either the domain
of the \code{\linkS4class{NMRSignalModel}}, or to 3 fwhm around the
\code{\linkS4class{NMRSignal1D}}'s outer peaks

For \code{\linkS4class{NMRSignalModel}} and list(x,y), the ppm resolution
is read from the argument. For \code{\linkS4class{NMRSignal1D}},
a resolution of 0.00023 ppm is used.
}
\section{Functions}{
\itemize{
\item \code{integral(nmr = NMRSignal1D, method = character)}: NMRSignal1D method

\item \code{integral(nmr = NMRSignalModel, method = character)}: NMRSignalModel method

\item \code{integral(nmr = NMRSignal1D, method = `function`)}: NMRSignal1D method with custom integration function

\item \code{integral(nmr = NMRSignalModel, method = `function`)}: NMRSignalModel method with custom integration function

\item \code{integral(nmr = list, method = character)}: x,y data series method

\item \code{integral(nmr = list, method = `function`)}: x,y data series method with custom integration function

}}
