% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lorentzian}
\alias{lorentzian}
\title{Lorentzian function
Interpolates a lorentzian with the given parameters on the given points.}
\usage{
lorentzian(x, mean = 0, max = 1, fwhm = 1)
}
\arguments{
\item{x}{numeric, points to be interpolated}

\item{mean}{numeric, mean(s) of the lorentzian(s)}

\item{max}{numeric, maxima of the lorentzian(s)}

\item{fwhm}{numeric, full-width-at-half-max(s) of the lorentzians(s)}
}
\value{
numeric, vector of interpolated values
}
\description{
Lorentzian function
Interpolates a lorentzian with the given parameters on the given points.
}
\examples{
lorentzian(1:20,max=1,mean=0)
}
