% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr.spectra.processing-package.R
\docType{package}
\name{nmr.spectra.processing-package}
\alias{nmr.spectra.processing}
\alias{nmr.spectra.processing-package}
\title{nmr.spectra.processing: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
