% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmrium.R
\name{nmrium}
\alias{nmrium}
\title{Visualize spectra in nmrium}
\usage{
nmrium(
  ppm,
  yr,
  yi = NULL,
  roi = range(ppm),
  cache = file.path(getwd(), "nmrium"),
  metadata = NULL,
  ...
)
}
\arguments{
\item{ppm}{numeric, chemical shift scale}

\item{yr, }{matrix, numeric, nmr intensities, real part, samples in rows}

\item{yi, }{matrix, numeric, nmr intensities, imaginary part, samples in rows
Default: NULL}

\item{roi}{numeric, optional. Upper and lower limit of the region of interest
to be cropped. Default \code{range(ppm)}}

\item{cache, }{directory to cache the data for visualization. A http server
will be setup in this directory. Default \code{getwd()}}

\item{metadata, }{data.frame, optional, additional experimental data to be
exported. Samples in rows.}

\item{..., }{additional arguments for server initialization and data export
(see \code{hastaLaVista2::exportReIm})}
}
\description{
Quality of life wrapper to \code{hastsaLaVista2} utils
}
\details{
R is a powerful software for data analysis, however it is not so
good at interactive plotting (we don't like plotly). This function interfaces
with nmrium to allow interactive spectra visualization (zoom etc.) and processing
(peak picking, integration, phase correction, assignment) from the web browser.
}
