% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noiseLevel.R
\name{noiseLevel}
\alias{noiseLevel}
\title{Estimates the noise level of the spectrum}
\usage{
noiseLevel(ppm, y, level = 0.99, rOref = c(9.8, 10))
}
\arguments{
\item{ppm}{numeric, chemical shift scale of the spectrum.}

\item{y}{numeric, spectrum intensities.}

\item{level}{numeric, quantile to estimate noise, specified as a probability
as in \code{\link[stats]{quantile}}. Default: 0.99.}

\item{rOref}{numeric, limits of the Region of Reference used to estimate
noise. Default: 9.8 - 10 ppm.}
}
\value{
numeric, estimation of the spectrum's level of noise.
}
\description{
Estimates the noise level of the spectrum
}
\details{
Takes a 'blank' region of the spectrum as reference and
estimates noise as the \code{level}-probability quantile of intensities
in that region.
}
