% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Spectrum normalization}
\usage{
normalize(X, method = pqn, ...)
}
\arguments{
\item{X}{numeric vector (single spectrum) or matrix (multiple spectra, samples
in) with NMR intensities}

\item{method, }{function to be used for normalization. Defaults to \code{\link{pqn}}}

\item{..., }{additional parameters to method}
}
\value{
normalized spectra matrix.
}
\description{
Spectra normalization
}
\details{
This is mostly just a convenience wrapper to compute X / method(X)
under the expectation that, for a n x m matrix X, method(X) returns a vector
of n (inverse) normalization factors. As added value, it parses certain values
of the \code{method} argument. Namely: \code{method=max} is interpreted as
max height normalization and \code{method=sum} is interpreted as total
area normalization.
}
