% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizeSignal}
\alias{normalizeSignal}
\title{Normalize a \code{\linkS4class{NMRSignal1D}}}
\usage{
normalizeSignal(signal)
}
\arguments{
\item{signal}{a \code{\linkS4class{NMRSignal1D}}}
}
\value{
scaled \code{\linkS4class{NMRSignal1D}}
}
\description{
Scales signal height to a maximum of 1
}
