% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad.R
\name{pad}
\alias{pad}
\title{Pads a series on either extreme with the number of specified "blank" points}
\usage{
pad(
  x,
  n,
  side = c(-1, 0, 1)[3],
  method = c("zeroes", "circular", "sampling")[1],
  from = as.integer(length(x) * 14/15):length(x)
)
}
\arguments{
\item{x}{numeric vector, series to be padded}

\item{n}{numeric, number of points to be added}

\item{side}{numeric, whether to pad at the start (-1), at the end (1, default)
or at both extremes (0)}

\item{method}{character, padding method. "zeroes" (default) pads with 0.
"circular" pads the end with points from the start and the start with points
from the end. "sampling" pads with points sampled from the input series}

\item{from}{logical or integer, optional. Filter selecting the region from x
to be used in the "sampling" method. Default: the last 1/15th points}
}
\value{
the padded series
}
\description{
Pads a series on either extreme with the number of specified "blank" points
}
