% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseCorrection.R
\name{phaseCorrection}
\alias{phaseCorrection}
\title{Apply a phase correction to a complex NMR spectrum. phi1 is specified at index=0}
\usage{
phaseCorrection(yr, yi, phi0, phi1, reverse = FALSE)
}
\arguments{
\item{yr}{list, spectrum intensities, real part}

\item{yi}{list, spectrum intensities, imaginary part}

\item{phi0}{numeric, phase 0 order correction.}

\item{phi1}{numeric, phase 1 order correction.}

\item{reverse}{boolean, TRUE if the chemical shift scale is inverted
(decreases from left to right). Default=FALSE}
}
\value{
list, A list with 2 elements corresponding to the corrected spectrum (yr, yi)
}
\description{
Apply a phase correction to a complex NMR spectrum. phi1 is specified at index=0
}
