% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{pqn}
\alias{pqn}
\title{Computes the PQN estimation of the dilution factor for a spectrum or matrix of spectra}
\usage{
pqn(X, ref = median, of = TRUE)
}
\arguments{
\item{X}{numeric vector or matrix, input NMR}

\item{ref}{function or numeric vector, reference for the PQN. Can be provided
as either a function to be called on the spectra matrix to compute the reference
spectrum, or a vector giving the reference spectrum directly. Defaults to the
median spectrum.}

\item{of}{logical or integer, a filter to select the rows of the spectra matrix
to be used in the calculation of the reference.}
}
\value{
numeric vector with the PQN dilution factors. Note that to apply the
correction each spectrum must be \emph{divided} by the corresponding dilution factor.
}
\description{
Computes the PQN estimation of the dilution factor for a spectrum or matrix of spectra
}
\details{
See doi:10.1021/ac051632c
}
