% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pseudoVoigt}
\alias{pseudoVoigt}
\title{Pseudo-Voigt function}
\usage{
pseudoVoigt(x, mean = 0, max = 1, fwhm = 1, mu = 0)
}
\arguments{
\item{x}{numeric points to be interpolated}

\item{mean}{numeric, mean of the distribution}

\item{max}{numeric, maxima of the lorentzian(s)}

\item{fwhm}{numeric, full width at half max of the distribution}

\item{mu}{numeric, <= 1. The mu parameter of the pseudo-Voigt approximation.
The coefficients of the linear combination are mu for the lorentzian and
1 - mu for the gaussian}
}
\value{
numeric, vector of interpolated values
}
\description{
Interpolates a Voigt function, using the pseudo-Voigt approximation,
on the given points with the given parameters.
The pseudo-Voigt approximation approximates a Voigt function as a linear
combination of a lorentzian and a gaussian.
}
\examples{
pseudoVoigt(1:100,max=1,mean=0,mu=0.8)
}
