% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftSeries.R
\name{shiftSeries}
\alias{shiftSeries}
\title{Shifts a series by the given value}
\usage{
shiftSeries(
  x,
  shift,
  padding = "sampling",
  from = as.integer(length(x) * 14/15):length(x)
)
}
\arguments{
\item{x}{numeric, the series to be shifted}

\item{shift}{numeric, the amount to be shifted. Negative shifts to the left.}

\item{padding}{character, the method to be used to fill the empty extremes of
the shifted series. See \code{\link{pad}} for details. Default: "zeroes".}

\item{from}{logical or integer, optional. Filter selecting the region from x
to be used in the "sampling" method, see \code{\link{pad}} for details.
Default: the last 1/15th points.}
}
\value{
numeric, the shifted series
}
\description{
Shifts a series by the given value
}
