% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftSeries.R
\name{shiftSpectra}
\alias{shiftSpectra}
\title{Shifts spectra by the given frequency}
\usage{
shiftSpectra(ppm, Y, shift, hertz = FALSE, SF = 600, interpol = "spline")
}
\arguments{
\item{ppm}{numeric, chemical shift scale}

\item{Y}{numeric matrix or vector, NMR intensities, spectra in rows}

\item{shift}{numeric, the frequency to shift each spectrum by. Recycled as
necessary, see details}

\item{hertz}{logical. If TRUE, \code{shift} is interpreted as an
absolute frequency shift in hertz. If FALSE (default), it is interpreted as
a relative ppm chemical shift}

\item{SF}{numeric, field strength in megahertz for each spectrum. Used to
convert \code{shift} to ppm units if hertz=TRUE. Recycled as necessary, see
details}

\item{interpol}{character, the interpolation method to use, see details.
See  \code{\link[signal]{interp1}} for accepted values. The default "spline"
should works best in most cases.}
}
\value{
The shifted spectra
}
\description{
Shifts spectra by the given frequency
}
\details{
Works by adding shift to ppm and then interpolating Y at ppm assuming
Y aligned with the shifted ppm scale. Arguments \code{shift} and \code{SF} are
recycled to match the number of rows of Y. No warnings are raised, be responsible.
}
