% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{signalDomain}
\alias{signalDomain}
\title{Calculate the domain of a \code{\linkS4class{NMRSignal1D}}}
\usage{
signalDomain(signal, n = 3)
}
\arguments{
\item{signal, }{\code{\linkS4class{NMRSignal1D}}}

\item{n}{numeric, linewidth multiplier used to offset the outer peaks}
}
\value{
numeric, the extremes of the signal's domain
}
\description{
Estimates the domain of the signal as the coordinates of the outer peaks
offset by \emph{n} linewidths (strictly: \emph{n} times full-width at half-max)
}
