% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signalsToY.R
\name{signalToY}
\alias{signalToY}
\title{Interpolates a \code{\linkS4class{NMRSignal1D}} on the given set of points}
\usage{
signalToY(signal, ppm)
}
\arguments{
\item{signal, }{\code{\linkS4class{NMRSignal1D}}, signal to interpolate}

\item{ppm, }{chemical shift values to interpolate}
}
\value{
numeric vector of interpolated intensities
}
\description{
Interpolates a \code{\linkS4class{NMRSignal1D}} on the given set of points
}
