% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSpectra.R
\name{alignSeries}
\alias{alignSeries}
\title{Aligns series to a reference
Computes the shift that maximizes each series' cross-correlation to the}
\usage{
alignSeries(Y, ref = "median", shifts = FALSE, ...)
}
\arguments{
\item{Y}{matrix, series in rows}

\item{ref}{character, numeric or function. Specifies the reference for
alignment. May be a series of the same length as the input series, the index
of the row of Y to be used as reference, a function to compute the
reference from the input, or a character representing an implemented
reference. Valid character references are 'median' (default: reference is
the median of the input) and 'mean'. If ref is a function, it run on Y
and the result is used as reference.}

\item{shifts}{logical. If TRUE, returns the shifts; if FALSE, returns the
shifted series. Default: FALSE}

\item{...}{additional parameters}
}
\value{
either a numeric vector of shifts or a matrix with the shifted
series in its rows
}
\description{
Aligns series to a reference
Computes the shift that maximizes each series' cross-correlation to the
}
