% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrateSpectra.R
\name{calibrateSpectra}
\alias{calibrateSpectra}
\title{Spectra calibration}
\usage{
calibrateSpectra(
  ppm,
  Y,
  ref = c("tsp", "glucose", "alanine", "NMRSignal1D see documentation")[1],
  frequency = 600,
  maxShift = 1/3,
  threshold = 0.2,
  rOref,
  cshift,
  j,
  padding = c("zeroes", "circular", "sampling")[1],
  from = as.integer(length(ppm) * 14/15):length(ppm),
  ...
)
}
\arguments{
\item{ppm}{numeric, spectra chemical shift scale}

\item{Y}{matrix, intensities, spectra in rows}

\item{ref}{character or \code{\linkS4class{NMRSignal1D}}, the reference signal
for calibration. Either "glucose", "alanine", "tsp", or a
\code{\linkS4class{NMRSignal1D}}. "tsp" calibrates to a singlet at 0 ppm. "alanine"
calibrates to a doublet at 1.48 ppm with j=7.26. "glucose" aligns to a doublet
at 5.223 with j=3.63.}

\item{frequency}{numeric, the spectrometer frequency in MHz. Only used for
character \code{ref}. Default: 600}

\item{maxShift}{numeric, the maximum allowed shift as a fraction of the rOref,
see \code{\link[stats]{ccf}}}

\item{threshold}{numeric, minimum cross-correlation requiered for alignment.
See \code{\link{alignSeries}}}

\item{rOref}{numeric, optional. Limits of the Region of Reference within
which the reference signal for calibration will be aligned. Use if the
calibration signal is very shifted, or if you need to avoid interfering signals.
Otherwise, the default should be fine.}

\item{cshift, }{numeric, optional. Chemical shift of the reference. If
provided, it replaces the internal model's chemical shift. Use if you want
to position your signal differently from the standard model (see \code{ref})}

\item{j}{numeric, optional. Coupling constant of the reference. If provided,
it  replaces the internal model's coupling constant. Use it to fine tune
calibration to alanine and glucose, which may not be accurate due to
fluctuations in the observed \emph{j}.}

\item{padding}{character, the method to be used to fill the empty extremes of
the shifted spectra. See \code{\link{pad}} for details. Default: "zeroes".}

\item{from}{numeric, optional. Filter for the spectral region to be used for
padding after the spectrum is shifted in the "sampling" method. See
\code{\link{pad}} for details. Default: last 1/15th points}

\item{...}{additional arguments for \code{\link{alignSeries}}, see \strong{Details}}
}
\value{
calibrated spectra matrix
}
\description{
Calibrates spectra
}
\details{
Interpolates the reference signal with \code{\link{signalToY}} to
obtain a trace. Then, it aligns each row of the spectra matrix to this trace,
within the \code{rOref},  using \code{\link{alignSeries}}. Passing a NMRSignal1D
to \code{ref} allows to calibrate to any custom signal; but if you are aligning
to a custom doublet or singlet, it's more practical to set \code{ref} to either
"tsp" (for singlets) or "alanine" (for doublets) and adjust the signal parameters
through the arguments \code{cshift} and \code{j}. This way \code{calibrateSpectra}
handles the annoying task of translating the signal parameters into a NMRSignal1D
object.
}
