% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getI}
\alias{getI}
\title{Get the index of a chemical shift}
\usage{
getI(ppm, ...)
}
\arguments{
\item{ppm, }{numeric, chemical shift scale}

\item{..., }{numeric, chemical shift values to be found}
}
\value{
integer, the index of the element of ppm that is closest to v
}
\description{
Seeks the element of the chemical shift scale closest to the given value.
Useful e.g. to get the approximate intensity of spectra at a given c. shift.
Quality of life function to save you a few key strokes and some neural pulses.
Questionable value.
}
