% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmr.spectra.processing-package.R
\docType{package}
\name{nmr.spectra.processing-package}
\alias{nmr.spectra.processing}
\alias{nmr.spectra.processing-package}
\title{nmr.spectra.processing: Process Fourier-transformed NMR spectra}
\description{
Collection of tools for processing Fourier-transformed NMR spectra: baseline correction, shifting, alignment, calibration to a reference resonance, peak integration and more. Also packs some utils to help navigate NMR data without leaving R.
}
\author{
\strong{Maintainer}: Andrés Bernal \email{AndresFernando.BernalEscobar@murdoch.edu.au} (\href{https://orcid.org/0000-0003-1668-7225}{ORCID})

}
\keyword{internal}
