% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sannotate.R
\name{sannotate}
\alias{sannotate}
\title{Basic tool to annotate signals in a 1D NMR spectrum}
\usage{
sannotate(
  ppm,
  y,
  ann,
  roi,
  tcol = "black",
  lcol = "black",
  window = 0.002,
  ylim = "auto",
  add = FALSE,
  adj = c(0.5, -0.5),
  delta = 0,
  epsilon = 0.02,
  ...
)
}
\arguments{
\item{ppm}{numeric, chemical shift scale}

\item{y}{numeric, NMR intensities}

\item{ann}{list, signals to annotate. Each list element is a numeric vector
with the ppm coordinates of the signal's peaks. Signal labels are to be passed
as the list's names}

\item{roi}{numeric, optional. Upper and lower limit of the Region of Interest
to be plotted}

\item{tcol, }{text color for annotation}

\item{lcol, }{line color for annotation}

\item{window, }{numeric, ppm window for automatic annotation position adjustement.
Default 0.002. See details.}

\item{ylim, }{by default ("auto") plotting area is adjusted to make sure all
labels fit. You may override this behavior by providing explicit limits
as in \code{\link[base]{plot}}}

\item{add, }{logic, should the annotation be added to the active plot (TRUE) or a
new plot created (FALSE)? This is useful to manually position each individual label
for optimal visualization, to pass different graphic parameters (e.g. lwd) for
spectrum trace and for annotation, or to keep a cleaner code in extensive annotations.
Default FALSE.}

\item{adj, }{numeric, positioning of labels, see \code{\link[graphics]{par}}}

\item{delta, }{horizontal shift, in chemical shift units, of the text labels relative
to the peak positions}

\item{epsilon, }{numeric, scaling factor for marker length, scales it
proportionally to the plot window. More precisetly, this controls the vertical
spacing between the top peak and the text label, so the actual length
of the marker depends on other factors such as \code{delta} positioning. Default 0.02}

\item{...}{additional graphical parameters, see \code{\link[graphics]{par}},
see \code{\link{smatplot}}.}
}
\description{
Plots a spectrum with signals labeled
}
\details{
Annotations are re positioned to the nearest maximum intensity within
\code{window} of the given chemical shifts. You may compare the default
with  \code{window=0} to see the effect. Since annotations need the spectrum
for adjustment, arguments \code{ppm} and \code{y} are mandatory even if
\code{add=TRUE}. For a similar reason, you should pass the original
\code{ylim} to further calls using \code{add=TRUE}.
}
