% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{smatplot}
\alias{smatplot}
\title{Fast spectra overlay plot}
\usage{
smatplot(
  ppm,
  y,
  by,
  roi,
  type = "l",
  lty = 1,
  reverse = TRUE,
  legend,
  label,
  palette = c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", "#FF7F00", "#FFFF33",
    "#A65628", "#F781BF", "#999999"),
  ...
)
}
\arguments{
\item{ppm}{numeric, spectra ppm scale}

\item{y}{numeric or matrix, intensities, spectra in rows}

\item{by}{numeric, optional, number of spectra to be overlayed on each plot}

\item{roi}{numeric, optional, limits of the Region of Interest to be plotted.
Defaults to the range of \code{ppm}.}

\item{type}{optional, defaults to "l" (lines), default recommended for speed}

\item{lty}{optional, defaults to 1 (continuous line)}

\item{reverse}{logic, optional, if TRUE (default) the x scale of the plot
increases from right to left as it's customary in NMR spectroscopy}

\item{legend, }{optional, position of the legend as specified in
\code{\link[graphics]{legend}}}

\item{label, }{character, optional, series labels}

\item{palette}{vector of colors, equivalent to matplot(col). The default is
Set1 copied from RColorBrewer}

\item{..., }{additional arguments to be passed to \code{\link[graphics]{matplot}}}
}
\description{
Wrapper to \code{\link[graphics]{matplot}} for plotting spectra. Crude
compared to \code{ggplot2} but faster.
but faster.
}
